requirements.txt
tensorflow
numpy
pillow
scikit-image
scipy
augmentor
matplotlib
pandas
seaborn